Adding segment header and Compactor

+++ logstore/src/main/java/io/joshworks/fstore/log/appender/level/Levels.java
@@ -1,6 +1,6 @@
 package io.joshworks.fstore.log.appender.level;
 
-import io.joshworks.fstore.log.Log;
+import io.joshworks.fstore.log.segment.Log;
 import io.joshworks.fstore.log.appender.Order;
 
 import java.util.ArrayList;
@@ -8,6 +8,7 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class Levels<T, L extends Log<T>> {
@@ -22,7 +23,7 @@ public class Levels<T, L extends Log<T>> {
         this.flattened = flatten();
     }
 
-    public List<L> segments(int level) {
+    public synchronized List<L> segments(int level) {
         return items.get(level);
     }
 
@@ -41,13 +42,13 @@ public class Levels<T, L extends Log<T>> {
         return new Levels<>(maxItemsPerLevel, segments);
     }
 
-    public List<List<String>> segmentNames() {
+    public synchronized List<List<String>> segmentNames() {
         return items.stream()
                 .map(segs -> segs.stream().map(Log::name).collect(Collectors.toList()))
                 .collect(Collectors.toList());
     }
 
-    public void promoteLevelZero(L newLevelZero) {
+    public synchronized void promoteLevelZero(L newLevelZero) {
         List<L> levelZero = items.get(0);
         if (levelZero == null) {
             throw new IllegalStateException("No segment available in level zero");
@@ -73,14 +74,14 @@ public class Levels<T, L extends Log<T>> {
     }
 
     private LinkedList<L> flatten() {
-        return items.stream().flatMap(Collection::stream).collect(Collectors.toCollection(LinkedList::new));
+        return new ArrayList<>(items).stream().flatMap(Collection::stream).collect(Collectors.toCollection(LinkedList::new));
     }
 
     public int numSegments() {
         return flattened.size();
     }
 
-    public int size(int level) {
+    public synchronized int size(int level) {
         if(level < 0) {
             throw new IllegalArgumentException("Level must be at least zero");
         }
@@ -97,7 +98,7 @@ public class Levels<T, L extends Log<T>> {
         return items.get(level).size() >= maxItemsPerLevel;
     }
 
-    public void add(int level, L segment) {
+    public synchronized void add(int level, L segment) {
         if (depth() <= level) {
             items.add(new ArrayList<>(maxItemsPerLevel));
         }
@@ -106,22 +107,32 @@ public class Levels<T, L extends Log<T>> {
         this.flattened = flatten();
     }
 
-    public List<L> segmentsForCompaction(int level) {
+    public synchronized List<L> segmentsForCompaction(int level) {
         List<L> segments = segments(level);
-        return segments.size() > maxItemsPerLevel ? segments.subList(0, maxItemsPerLevel) : segments;
+        return segments.size() > maxItemsPerLevel ? new ArrayList<>(segments.subList(0, maxItemsPerLevel)) : new ArrayList<>(segments);
     }
 
-    public void removeSegmentsFromCompaction(int level) {
-        Iterator<L> segments = segmentsForCompaction(level).iterator();
-        while(segments.hasNext()) {
-            L segment = segments.next();
-            segment.delete();
-            segments.remove();
+    public synchronized void removeSegmentsFromCompaction(List<L> levelSegments) {
+
+        Set<String> namesToBeDeleted = levelSegments.stream().map(Log::name).collect(Collectors.toSet());
+
+        for (String segmentName : namesToBeDeleted) {
+            for (List<L> level : items) {
+                Iterator<L> iterator = level.iterator();
+                while(iterator.hasNext()) {
+                    L segment = iterator.next();
+                    if(segment.name().equals(segmentName)) {
+                        segment.delete();
+                        iterator.remove();
+                    }
+                }
+            }
         }
+        this.flattened = flatten();
     }
 
     //TODO safely swap segments
-    public void deleteSegments(int level) {
+    public synchronized void deleteSegments(int level) {
         if (depth() < level) {
             throw new IllegalArgumentException("No such level " + level + ", current depth: " + depth());
         }
